Download confirmation for each courses.
When we want to download all the courses using "php codecourse download" without specifying course name, we want to download all the courses except few ones. The original command will download all the courses anyway. What about asking user for each course if they want to download the course or not ?

Initially when we run "php codecourse download" in CMD, this will ask if you want confirmation for each course or not. If you don't need then it will download courses anyway.